pyproject.toml
[project]
name = "dj-store"
version = "0.1.0"
description = "Demo/PoC for vape shop store manager"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argcomplete>=3.6.2",
    "django>=5.2.4",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
]
